Migrate Quadrangulation
Quadragnulation will now happen in the primitive designated to handle
it, everything else will see the polygon as a border set. There are
still a few odd cases and potential for non-resolving loops within the
quadrangulation method, involving more than 2 points on a shared line.
Naive Displacements are now generated by the solid primitive, when
passed proper information for it.